Add a module file
